requirements.txt
-i https://pypi.org/simple
click==8.1.7
dataclasses==0.6
faker==30.0.0; python_version >= '3.8'
psycopg2-binary==2.9.9; python_version >= '3.7'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv[cli]==1.0.1; python_version >= '3.8'
pytz==2024.2
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
